Wheel of fortune - WheelGame - Reorganizacja - WheelGame rozbite na moduły. Zmiana przy wywoływaniu możliwości zgadnięcia całego hasła.
