+++ src/main/java/com/sda/games/wheelOfFortune/csv/WOFCategories.csv
@@ -5,6 +5,6 @@
 5,Chemia
 6,Fizyka
 7,Religie
-8,Mechanika
-9,Informatyka
-10,Anatomia
+8,Informatyka
+9,Anatomia
+10,Sport
