+++ src/main/java/com/sda/games/wheelOfFortune/methods/Administration.java
@@ -1,0 +1,5 @@
+package com.sda.games.wheelOfFortune.methods;
+
+public class Administration {
+    //tu będą opcje dodawania, usuwania słów i kategorii
+}
